+++ src/Factory/Xml/FeedResponseFactory.php
@@ -4,35 +4,40 @@ declare(strict_types=1);
 
 namespace Linio\SellerCenter\Factory\Xml;
 
-use Linio\SellerCenter\Exception\InvalidXmlStructureException;
 use Linio\SellerCenter\Response\FeedResponse;
+use Linio\SellerCenter\Validator\XmlStructureValidator;
 use SimpleXMLElement;
 
 class FeedResponseFactory
 {
+    private const XML_MODEL = 'Feed';
+    private const REQUIRED_FIELDS = [
+        'RequestId',
+        'RequestAction',
+        'ResponseType',
+        'Timestamp',
+    ];
+
     public static function make(SimpleXMLElement $xml): FeedResponse
     {
-        if (!property_exists($xml, 'RequestId')) {
-            throw new InvalidXmlStructureException('Feed', 'RequestId');
+        XmlStructureValidator::validateStructure($xml, self::XML_MODEL, self::REQUIRED_FIELDS);
+
+        $requestParameters = [];
+        if (property_exists($xml, 'RequestParameters')) {
+            foreach ($xml->RequestParameters->children() as $item) {
+                $requestParameters[$item->getName()] = (string) $item;
+            }
         }
 
-        if (!property_exists($xml, 'RequestAction')) {
-            throw new InvalidXmlStructureException('Feed', 'RequestAction');
-        }
-
-        if (!property_exists($xml, 'ResponseType')) {
-            throw new InvalidXmlStructureException('Feed', 'ResponseType');
-        }
-
-        if (!property_exists($xml, 'Timestamp')) {
-            throw new InvalidXmlStructureException('Feed', 'Timestamp');
-        }
+        $requestId = !empty($xml->RequestId) ?
+            (string) $xml->RequestId : null;
 
         return new FeedResponse(
-            (string) $xml->RequestId,
+            $requestId,
             (string) $xml->RequestAction,
             (string) $xml->ResponseType,
-            (string) $xml->Timestamp
+            (string) $xml->Timestamp,
+            $requestParameters
         );
     }
 }